commiting mix.lock
The Elixir community seems to be split on when to version the mix.lock
file. Yehuda wrote about Gemfile.lock that it shouldn't necessarily be
locked by libraries but should be locked by individual deploys. Since
this is a personal project that won't "deploy" anywhere it really
doesn't matter. I'll defualt to verisoning the lock until I learn more
about how that is painful.